+++ NEWS.md
@@ -2,6 +2,8 @@
 ## nlrx 0.4.2
 * hotfix for another dependency on external files in nldoc roxygen examples
 * small bugfix in analyze_morris: A warning is no thrown if NA are present in the simulation data
+* bugfix in random seed generator
+* added option to run_nl_one that allows to store results as rds files
 
 ## nlrx 0.4.1
 * fixed dependency on external file source in nldoc automated tests